Initial commit
This is a GitOps action for Tailscale users. This enables users to set
up their Tailscale ACLs in a GitHub repository and then have changes be
pushed out by CI. This also allows changes to be reviewed in CI and have
CI fail if ACL tests fail.

More details to come.

Signed-off-by: Xe <xe@tailscale.com>