readme: fix actions/cache usage
Explicitly name only the version-cache.json file to be cached, so as to
not have the cache overwrite the policy.hujson file.

Create a new cache key for each successful run, and restore the latest
key by ordered prefix match so as to avoid stale cache reads.

Fixes #7
Fixes #8
Fixes #11