action.yml: add separate steps for API key and OAuth flows
Add separate steps for invocations of gitops-pusher for using API key
and OAuth. These steps are triggered based on whether the inputs for the
API key or OAuth secret are non-empty respectively. Add a check for
whether both an API key and OAuth secret have been specified and error out
in this case. Changing the logic in gitops-pusher to only attempt OAuth
authentication when the provided secret is non-empty is likely also a
desired change on top of this.

Signed-off-by: Mario Minardi <mario@tailscale.com>