action.yml: recombine gitops-pusher calls
We previously split the gitops-pusher steps based on which auth
variables were set due to a bug in the gitops-pusher tool itself.  That
as fixed in tailscale/tailscale#11025, so we can now recombine these.

Update to latest tagged version of tailscale to pull in the fix, and
bump the go version while we're at it.

Co-authored-by: Mario Minardi <mario@tailscale.com>
Signed-off-by: Will Norris <will@tailscale.com>